+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Fixed a fault that allowed events to be added in the Weekly view when manual event adding/editing is disabled.
 - Added new event "onNotificationClicked", which is triggered when an event notification is clicked.
 - Added new option "allowEventScrollingOnMainDisplay", which allows days in the main display to be scrolled (see documentation).
+- Added new public function addHoliday, which will add a new holiday and re-draw the calender.
 - 
 
 ### Version 0.9.9: